+++ src/Abp/Framework/Abp/Application/Session/NullAbpSession.cs
@@ -11,7 +11,7 @@
         public static NullAbpSession Instance { get { return SingletonInstance; } }
         private static readonly NullAbpSession SingletonInstance = new NullAbpSession();
         
-        public int? UserId { get { return null; } }
+        public long? UserId { get { return null; } }
         
         public int? TenantId { get { return null; } }
