+++ src/Abp/Framework/Abp/Domain/Entities/Auditing/ICreationAudited.cs
@@ -2,13 +2,13 @@ namespace Abp.Domain.Entities.Auditing
 {
     /// <summary>
     /// This interface is implemented by entities which's creation informations (who and when created) must be stored.
-    /// Creation time and creator user are automatically set when saving <see cref="IEntity"/> to database.
+    /// Creation time and creator user are automatically set when saving <see cref="Entity"/> to database.
     /// </summary>
     public interface ICreationAudited : IHasCreationTime
     {
         /// <summary>
         /// Creator of this entity.
         /// </summary>
-        int? CreatorUserId { get; set; }
+        long? CreatorUserId { get; set; }
     }
 }
