+++ src/Taskever/Taskever.Application/Users/Dto/ConfirmEmailInput.cs
@@ -4,8 +4,8 @@ namespace Abp.Users.Dto
 {
     public class ConfirmEmailInput
     {
-        [Range(1, int.MaxValue)]
-        public int UserId { get; set; }
+        [Range(1, long.MaxValue)]
+        public long UserId { get; set; }
 
         [Required]
         public string ConfirmationCode { get; set; }
