+++ src/Taskever/Taskever/Activities/IUserFallowedActivityRepository.cs
@@ -5,6 +5,6 @@ namespace Taskever.Activities
 {
     public interface IUserFollowedActivityRepository :  IRepository<UserFollowedActivity, long>
     {
-        IList<UserFollowedActivity> Getactivities(int userId, bool? isActor, long beforeId, int maxResultCount);
+        IList<UserFollowedActivity> Getactivities(long userId, bool? isActor, long beforeId, int maxResultCount);
     }
 }
